+++ src/TrashLib/Services/Radarr/CustomFormat/RadarrGuideDataLister.cs
@@ -1,16 +1,17 @@
 using CliFx.Infrastructure;
 using JetBrains.Annotations;
+using MoreLinq;
 using TrashLib.Services.Radarr.CustomFormat.Guide;
 
 namespace TrashLib.Services.Radarr.CustomFormat;
 
 [UsedImplicitly]
-public class CustomFormatLister : ICustomFormatLister
+public class RadarrGuideDataLister : IRadarrGuideDataLister
 {
     private readonly IConsole _console;
     private readonly IRadarrGuideService _guide;
 
-    public CustomFormatLister(IConsole console, IRadarrGuideService guide)
+    public RadarrGuideDataLister(IConsole console, IRadarrGuideService guide)
     {
         _console = console;
         _guide = guide;
@@ -29,5 +30,18 @@ public class CustomFormatLister : ICustomFormatLister
         _console.Output.WriteLine(
             "\nThe above Custom Formats are in YAML format and ready to be copied & pasted " +
             "under the `trash_ids:` property.");
+    }
+
+    public void ListQualities()
+    {
+        _console.Output.WriteLine("\nList of Quality Definition types in the TRaSH Guides:\n");
+
+        _guide.GetQualities()
+            .Select(x => x.Type)
+            .ForEach(x => _console.Output.WriteLine($"  - {x}"));
+
+        _console.Output.WriteLine(
+            "\nThe above quality definition types can be used with the `quality_definition:` property in your " +
+            "recyclarr.yml file.");
     }
 }