+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Radarr: New `--list-qualities` argument to get a list of quality definition types from the guide.
+
 ### Changed
 
 - Quality definition data is now pulled from JSON files.